pyproject.toml
[project]
name = "search-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "black>=26.3.1",
    "isort>=8.0.1",
    "langchain>=1.2.15",
    "langchain-ollama>=1.0.1",
    "langchain-tavily>=0.2.17",
    "python-dotenv>=1.2.2",
    "tavily-python>=0.7.23",
]
